+++ src/Contract/Mutator.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Rougin\Weasley\Contract;
+
+/**
+ * Mutator Interface
+ *
+ * @package Weasley
+ * @author  Rougin Gutib <rougingutib@gmail.com>
+ */
+interface Mutator
+{
+    /**
+     * Mutates the contents of the result.
+     *
+     * @param  mixed $data
+     * @return mixed
+     */
+    public function mutate($data);
+}