+++ tests/Fixture/Controllers/UsersController.php
@@ -20,5 +20,10 @@ class UsersController extends RestfulController
     /**
      * @var string
      */
+    protected $transformer = 'Rougin\Weasley\Transformer\ApiTransformer';
+
+    /**
+     * @var string
+     */
     protected $validator = 'Rougin\Weasley\Fixture\Validators\UserValidator';
 }